requirements.txt
asgiref==3.2.10
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-toolbelt==0.0.1
djangorestframework==3.12.4
easy-mask==1.3.4
gunicorn==20.0.4
idna==2.10
Pillow==8.2.0
psycopg2==2.9.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.5
whitenoise==5.2.0
